refactoring and cleaning
better security measures
Producer / Series Actions are now better separated
fixed bugs in listing series episodes (<,> bug)
episodeStack now has an option to link to an pager if more results are
available
you can now browse all episodes of an series (with sorting)
more assertions for forms (can’t send an empty commend)
moved templates accordingly
simplified security roles and hierarchy
