+++ app/config/security.yml
@@ -5,6 +5,11 @@ security:
         ROLE_USER: ROLE_OKTOLAB_PRODUCER
         ROLE_OKTOLAB_PRODUCER: ROLE_OKTOLAB_USER
 
+# ROLE_ADMIN:            TZE BOSSA
+# ROLE_USER:             Okto Ppl
+# ROLE_OKTOLAB_PRODUCER: Okto Serie Ppl
+# ROLE_OKTOLAB_USER:     Registered User
+
     providers:
         bprs_applink:
             entity:
