+++ src/AppBundle/Resources/views/Series/episodesWithTagsAjax.html.twig
@@ -10,6 +10,6 @@
         </select>
 	</div>
     <div id="episode-tag-stack">
-        {% include 'AppBundle::Default/EpisodeStack.html.twig' with {'episodes': episodes} %}
+        {% include 'AppBundle::Default/EpisodeStack.html.twig' with {'episodes': episodes, 'collapse_link': path('oktothek_show_series_episodes', {'uniqID': series.uniqID})} %}
     </div>
 </section>
