+++ src/AppBundle/Resources/views/Series/seriesEpisodes.html.twig
@@ -1,0 +1,34 @@
+{% extends "BprsStyleBundle::layout.html.twig" %}
+
+{% block title %}{{ name|default('OKTO')~' | '~ 'oktothek_series_episodes_title'|trans }}{% endblock %}
+{% block nav_head%}
+    {% include 'BprsStyleBundle::menu.html.twig' with {'selected': '', 'dropdown': ''} %}
+{% endblock %}
+
+{% set body_class = "head-white-color" %}
+{% block body_main %}
+<div class="page-header">
+    <h1><a href="{{ path('oktothek_show_series', {'uniqID': series.uniqID}) }}">{{ series }}</a></h1>
+    <div class="dropdown">
+      <button class="btn btn-default dropdown-toggle" type="button" id="filter_menu" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
+        {{ 'oktothek.series_episodes_filter'|trans|raw }}
+      </button>
+      <ul class="dropdown-menu" aria-labelledby="filter_menu">
+        <li {% if episodes.isSorted('e.name') %}class="active"{% endif %}>{{ knp_pagination_sortable(episodes, 'oktothek.series_episodes_name_filter'|trans, 'e.name') }}</li>
+        <li {% if episodes.isSorted('e.onlineStart') %}class="active"{% endif %}>{{ knp_pagination_sortable(episodes, 'oktothek.series_episodes_onlineStart_filter'|trans, 'e.onlineStart') }}</li>
+      </ul>
+    </div>
+</div>
+    <div class="navigation">
+        {{ knp_pagination_render(episodes) }}
+    </div>
+    <div class="row">
+    {% for episode in episodes %}
+        {% include 'AppBundle:Default:EpisodeThumb.html.twig' %}
+    {% endfor %}
+    </div>
+    <div class="navigation">
+        {{ knp_pagination_render(episodes) }}
+    </div>
+
+{% endblock %}
